requirements/base.txt
Django==1.11.4
django-environ==0.4.0
django-filter==1.0.4
djangorestframework==3.6.3
six==1.10.0
psycopg2==2.6.1
django-cors-headers==2.1.0
iso8601==0.1.11
tqdm==4.8.3
requests==2.11.1
elasticsearch-dsl==5.1.0
azure==2.0.0rc6
django-taggit==0.21.2
pytz==2016.6.1
newrelic==2.68.0.50
django-anymail==0.6.1
django-nose==1.4.4
beautifulsoup4==4.4.1
rosette_api==1.7.0
nameparser==0.5.3
responsebot==0.4.0
